code cleanup/organization
