+++ server/routes/itemRoutes.js
@@ -7,10 +7,10 @@ const router = express.Router();
 const itemController = require("../controllers/itemController");
 
 router.get("/:sectionId/items", itemController.getItems); // GET api/items/:sectionId/items
-router.post("/:sectionId/items/:username", itemController.createItem); // POST api/items/:sectionId/items
+router.post("/:sectionId/items/:username", itemController.createItem); // POST api/items/:sectionId/items/:username
 router.put("/:sectionId/items/:username/order", itemController.updateItemOrder); // PUT api/items/:sectionId/items/:username/order
-router.put("/:sectionId/items/:itemId/:username/move", itemController.moveItem); // PUT api/items/:sectionId/items/:itemId/move
+router.put("/:sectionId/items/:itemId/:username/move", itemController.moveItem); // PUT api/items/:sectionId/items/:itemId/:username/move
 router.put("/:sectionId/items/:itemId", itemController.updateItem); // PUT api/items/:sectionId/items/:itemId
-router.delete("/:sectionId/items/:itemId/:username", itemController.deleteItem); // DELETE api/items/:sectionId/items/:itemId
+router.delete("/:sectionId/items/:itemId/:username", itemController.deleteItem); // DELETE api/items/:sectionId/items/:itemId/:username
 
 module.exports = router;
