+++ server/utils/constants.js
@@ -2,15 +2,16 @@ const roomActions = Object.freeze({
   LEAVE: "leave-room",
   JOIN: "join-room",
   USERS: "room-users",
+  DISCONNECTING: "disconnecting",
 });
 
-const COLORS = [
+const COLORS = [ // array of colors for users
   "#ff6b6b",
   "#ffd93d",
   "#6bcb77",
   "#4d96ff",
   "#a66cff",
-  "#ff6F91",
+  "#ff6f91",
   "#ff9671",
 ];
